refactor: move non generic frontend-utils code into their respective frontends
