+++ jplag.frontend-utils/pom.xml
@@ -10,10 +10,4 @@
 		<version>3.1.0-SNAPSHOT</version>
 	</parent>
 
-	<dependencies>
-		<dependency>
-			<groupId>antlr</groupId>
-			<artifactId>antlr</artifactId>
-		</dependency>
-	</dependencies>
 </project>
