+++ jplag.frontend.text/src/main/java/de/jplag/text/ParserToken.java
@@ -1,4 +1,4 @@
-package de.jplag;
+package de.jplag.text;
 
 import antlr.Token;
